isa: allow cell array of class names as input, plus remove __isa_parent__().
* ov-class.cc (F__isa_parent__): removed function which was only used in isa.m
and has now be removed. Replaced by new isa (implemented in C++).
  (Fisa): new function to replace isa.m and F__isa_parent__. Now supports a
cell array of strings with classnames. While the m file implementation was
kind of supporting this, it was doing so by accident and would not work when
checking for parent class, integer, numeric, or float.
* scripts/general/module.mk: remove isa.m from list of files.
* scripts/general/isa.m: remove file. This function was really small and half
of it (__isa_parent__) had to be implemented in C++ anyway. Implementing the
whole thing in C++ means one less undocumented, meant as private, function.
And any extra performance increase in a function called multiples times for
input check is welcome.