Overhaul functions in scripts/set directory.
* set.txi: Rewrite documentation for set functions.

* intersect.m: Rewrite docstring.  Use by_rows variable for code clarity.
Return output orientation which is compatible with Matlab.  Add %!tests for
output orientation and N-dimensional inputs.

* setdiff.m: Rewrite docstring.  Use by_rows variable for code clarity.
Rename output i to ia to clarify it is an index into the set a.
Return output orientation which is compatible with Matlab.  Add %!tests for
N-dimensional inputs.

* setxor.m: Rewrite docstring.  Use by_rows variable for code clarity.
Return output orientation which is compatible with Matlab.  Add %!tests for
output orientation and N-dimensional inputs.

* union.m: Rewrite docstring.  Use by_rows variable for code clarity.
Return output orientation which is compatible with Matlab.  Add %!tests for
output orientation and N-dimensional inputs.  Add %!tests for validsetargs
which are common to all set functions.

* unique.m: Rewrite docstring.  Verify that input is numeric or cell array
of strings.  Avoid computing idx for optional i,j outputs unless required.
Add %!error tests for input validation.

* ismember.m: Rewrite docstring.  Use input variable 'a' instead of 'A' for
conformance with rest of set functions.  Rename output index variable to
s_idx for clarity that it is an index into the set s.

* powerset.m: Rewrite doctring.  Add input validation on nargin.  Add %!error
input validation tests.

* module.mk: Include validsetargs.m in build system.

* validsetargs.m: Function renamed from validargs which was too general.
* validargs.m: Function renamed to validsetargs.