build: build doc-cache and DOCSTRINGS files in the build dir
* libinterp/module.mk (libinterp/DOCSTRINGS): Build in the build directory,
allow vpath builds to find it in $(srcdir).
(install-built-in-docstrings): Use file in $(srcdir) if it doesn't exist in
the current directory.

* scripts/module.mk (scripts/DOCSTRINGS): Build in the build directory,
allow vpath builds to find it in $(srcdir).

* run-octave.in: Use doc-cache and DOCSTRINGS in $(srcdir) if they don't exist
in the current directory.

* doc/interpreter/mk_doc_cache.m: Add an argument for the source directory.
Apply it to each of the DOCSTRINGS input files if they don't exist in the
current directory.

* doc/interpreter/munge-texi.pl: Apply $top_srcdir to each of the DOCSTRINGS
input files if they don't exist in the current directory.

* doc/interpreter/module.mk (doc/interpreter/doc-cache): Call mk_doc_cache.m
with additional $(srcdir) argument.