Backed out changeset 08aa03dfb00e
This changeset can cause trouble like this:

>> gnuplot_binary("/usr/bin/gnuplot")
error: '__gnuplot_has_feature__' undefined near line 49 column 5
error: called from
    gnuplot_binary at line 49 column 5

if Octave is running in a build tree that is separate from the source
tree.  The problem happens because gnuplot_binary.m is generated and
so is stored in the build tree.  But the private functions like
__gnuplot_has_feature__ are in the source tree.  Since the private
directory where __gnuplot_has_feature__ is located is not a
subdirectory of the directory where gnuplot_binary is located, Octave
doesn't find it.

I also noticed another problem due to this changeset.  The test executes

  gnuplot_binary ("X")

and that now ends up calling __gnuplot_version__ which attempts to
execute the command "X --version".  On many systems, "X" is the X11
server.  So then I see errors about the X server failing to start
properly and Octave hangs.

It also seems bad that "gnuplot_binary (new_prog)" will now call
gnuplot_binary recursively.  It currently works because the persistent
value that stores the program name is set before calling
__gnuplot_version__, but that seems a bit fragile.

I'm also wondering how important it is to be able to change the
gnuplot binary during a single Octave session.  Is this something that
comes up frequently?