gsvd.cc: Update gsvd implementation in preparation for Matlab compatibility (bug #48807).
* gsvd.cc: Rewrite docstring to descripe the API that gsvd will implement.
Change input validation to accept an optional third argument.  Issue warning
if third argument is present as "economy-sized" gsvd is not yet implemented.
Change error message to be generic that A and B must be matrices, since it
is not possible to determine which input is at fault.
Use FIXME at locations that need attention.
Remove uses of deprecated error_state variable by using xmatrix_value to
simultaneously attempt a conversion, and issue an error if it fails.
Fix error message about Inf/NaN values which was not referencing the correct
input A.
Add simple BIST test verifying the decomposition and singular values are correct.
Tag all BIST tests with the bug number 48807.