make --enable-64 the default
* m4/octave_blas_f77_func.m4: Rename from ax_blas_f77_func.m4.
Determine integer size used by BLAS library instead of checking to see
whether it matches the default Fortran integer size.
* m4/module.mk: Update.
* acinclude.m4 (OCTAVE_CHECK_SIZEOF_FORTRAN_INTEGER): Determine size
of Fortran INTEGER instead of checking to see whether it matches what
we expect.
* configure.ac: Make --enable-64 the default, but change it to mean
setting the size of Octave's index type.  Check for the size of the
BLAS library integer type and ensure the Fortran compiler uses the
same size for integers.
* NEWS: Note change in --enable-64 default.
* doc/interpreter/install.txi: Update.