use pointer to scope instead of scope id
Eliminate symbol table scope ID and the global list of all scopes
indexed by numeric ID.  Function scope data is now only accessible
from the function itself, or by asking a scope for its parent scope
(if it exists).

The top-level and global scopes are now regular data members of the
symbol table class instead of being static.

Symbol table scopes are now created in the lexer when parsing a
function begins and stored in the function object when finishing the
construction of function object.  If an error occurs while parsing a
function, the list of any pending scopes is deleted.