fplot.m: Fix handling of non-vectorized input functions (bug #52673).
* fplot.m: Change documentation to note that performance is better with
vectorized functions, but that scalar functions are accepted.  If function is
determined to be scalar during input validation, use arrayfun to convert it to
a vectorized version.  Add eps to denominator of calculation of error so that 0
values do not result in NaN.  Add new BIST test for warning emitted when
function is not vectorized.