axes.m: Fix stacking of axes when legends, colorbars, annotations are present (bug #44410).
* axes.m: Change documentation to note that a "Cartesian" axes object is created.
Expand explanation of how axes objects are re-stacked to note that legends and colorbars
are also pulled to the top of the stack.  Rename variable "cf" to "hpar" because parent
of axes is not always a figure.  If new axes is being created there are no colorbars
or legends, so handle the single possibility remaining of restacking an annotation
layer directly in axes() function.  Add complex BIST test for restacking multiple
axes with multiple legends, colorbars, and annotation layer.
* axes.m (restack_axes): Rename variable "cf" to "hpar".  Rename variable "show" to
"shh" for clarity.  Add code to find any legend or colorbar axes associated with this
axes and restack them above the actual axes object so they are visible.