directly edit scalar numeric & logical values in cell arrays (bug #52988)
* variable-editor-model.cc (variable_editor::impl::cell): Don't
require subeditor for scalar numeric and logical values in cell
arrays.  Create label in same way for all objects.

* ov-cell.cc (octave_cell::edit_display): Correctly call edit_display
for individual scalar numeric and logical values instead of calling
edit_display recursively for all scalar objects.  Return dimensions
and type if element is not a scalar numeric or logical array.