publish.m: Improve code evaluation and add classdef support.
* scripts/miscellaneous/publish.m (eval_context): New subfunction.  The previous
code evaluation relied on storing variables to a temporary location, using
save() and load() to emulate a temporary function context.

Using eval_context() a solution inspired by the Octave manual has been chosen:
https://octave.org/doc/v4.4.0/Evaluation-in-a-Different-Context.html.
Instead of saving to a temporary hdd location, the data is saved in memory
inside a persistent variable that is cleared at the end of code publishing.

Somehow it is reverting the changes by Mike in cset ab1aa0e57b72, but with the
new approach the variable dependency is reduced to a single instance, which is
manageable.  Especially the "clear all"-bug (bug #51096) is no problem for this
new solution.

The pros of this solution are:
+ No dependency on temporary files
+ No file I/O for code evaluation
+ Support for not-storable data types (classdef objects, ...)