Have TexInfo errors result in warning delivered to user (patch #9464)
Previously, we called "makeinfo --force" so that we returned at least
partial output even from malformed TeXinfo.  However, this meant it was
not possible to determine that an error had occurred because "status"
was not set.  Instead we call makeinfo once, and only if there was an
error, we process a second time with --force.

* __makeinfo__.m: Call makeinfo first without "--force" option.
Check output status and re-invoke WITH "--force" if necessary.