Inverse of a sparse/diagonal singular matrix should be a sparse/diagonal matrix of Infs.
* liboctave/array/CSparse.cc(inverse): if the matrix is null, return an error.
If it is singular, return a sparse matrix of Infs.
* liboctave/array/dSparse.cc(inverse): if the matrix is null, return an error.
If it is singular, return a sparse matrix of Infs.
* liboctave/array/CDiagMatrix.cc(inverse): if the matrix is null, return
an error. If it is singular, return a diagonal matrix of Infs.
* liboctave/array/dDiagMatrix.cc(inverse): if the matrix is null, return
an error.  If it is singular, return a diagonal matrix of Infs.
* libinterp/corefcn/inv.cc: set rcond to zero for singular diagonal matrices.
Add tests for the above cases.
* test/diag-perm.tst: modify a test for the inversion of a singular diagonal
matrix.