make unimplmented message an error and append to undefined identifier error
* __unimplemented__.m: Throw error instead of issuing warning
message.  Update messages to be appropriate as additional info for
another error instead of as an error by itself.
* variables.h, variables.cc (Fmissing_function_hook): Update doc string.
(maybe_missing_function_hook): Call missing_function_hook with nargout
set to 1.  If result is a string, return it, otherwise return an empty
string.
* pt-id.cc (tree_identifier::eval_undefined_error): If
missing_function_hook returns a non-empty string, append it to the
undefined symbol error message.