move parse_fcn_file from interpreter member function to parser friend
In 24b7e6326e26 I moved parse_fcn_file to the interpreter class, but
on reflection it seems that this function is too low-level to be in
the public interface for the interpreter.  There are about 10
parameters and it is fairly confusing to use them properly.  It also
needs access to many internal parser and lexer settings, so making it
a friend of the parser makes more sense to me.

* parse.h, oct-parse.yy (parse_fcn_file): Now a friend of the
base_parser class instead of an interpreter class member function.
Change all uses.