improve exit behavior with qt graphics toolkit (bug #56952)
When I was able to reproduce it, the crash described in bug report
#56952 appeared to be happening when the Qt event loop was calling
QtHandles::qt_graphics_toolkit::create_object when the interpreter was
being deleted and the gh_manager object was already invalid, so
attempting to lock the mutex it contains lead to a segfault.

This change attempts to avoid that possibility by disabling the Qt
graphics toolkit when interpreter_qobject::execute is finished with
the interpreter.  We also allow the the qt_application object to
delete the interpreter.

* interpreter-qobject.h, interpreter-qobject.cc
(interpreter_qobject::graphics_fini): New function.
(interpreter_qobject::execute): Use an unwind_action to ensure that
graphics_fini is called after the interpreter is finished executing
commands.  Don't explicitly delete the interpreter.  Allow the
application context to do that job.