move some common actions into tree_walker base class
Instead of having most tree_walker class methods be pure virtual
functions, define them to be methods that walk the tree but do nothing
else.  This change allows us to eliminate some duplication in derived
classes, and makes it easier to define derived classes that need to
walk the parse tree but onny perform actions for a few parse tree
elements.

* pt-walk.h, pt-walk.cc (class tree_walker): Define most methods to
walk the tree.

* cdef-class.cc (class ctor_analyzer): Delete methods in base class
that are now handled by the base tree_walker class.
* pt-anon-scopes.h, pt-anon-scopes.cc (class tree_anon_scopes): Likewise.
* pt-bp.h, pt-bp.cc (tree_breakpoint): Likewise.
* pt-check.h, pt-check.cc (tree_checker): Likewise.
* pt-eval.h, pt-eval.cc (tree_evaluator): Likewise.
* pt-jit.cc (jit_convert): Likewise.
* pt-pr-code.h, pt-pr-code.cc (tree_print_code): Likewise.