use centralized file for copyright info for individual contributors
* COPYRIGHT.md: New file.

* In most other files, use "Copyright (C) YYYY-YYYY The Octave Project
Developers" instead of tracking individual names in separate source
files.  The motivation is to reduce the effort required to update the
notices each year.

Until now, the Octave source files contained copyright notices that
list individual contributors.  I adopted these file-scope copyright
notices because that is what everyone was doing 30 years ago in the
days before distributed version control systems. But now, with many
contributors and modern version control systems, having these
file-scope copyright notices causes trouble when we update copyright
years or refactor code.

Over time, the file-scope copyright notices may become outdated as new
contributions are made or code is moved from one file to
another. Sometimes people contribute significant patches but do not
add a line claiming copyright. Other times, people add a copyright
notice for their contribution but then a later refactoring moves part
or all of their contribution to another file and the notice is not
moved with the code. As a practical matter, moving such notices is
difficult -- determining what parts are due to a particular
contributor requires a time-consuming search through the project
history. Even managing the yearly update of copyright years is
problematic. We have some contributors who are no longer
living. Should we update the copyright dates for their contributions
when we release new versions? Probably not, but we do still want to
claim copyright for the project as a whole.

To minimize the difficulty of maintaining the copyright notices, I
would like to change Octave's sources to use what is described here:

https://softwarefreedom.org/resources/2012/ManagingCopyrightInformation.html

in the section "Maintaining centralized copyright notices":

  The centralized notice approach consolidates all copyright
  notices in a single location, usually a top-level file.
  This file should contain all of the copyright notices
  provided project contributors, unless the contribution was
  clearly insignificant. It may also credit -- without a copyright
  notice -- anyone who helped with the project but did not
  contribute code or other copyrighted material.

  This approach captures less information about contributions
  within individual files, recognizing that the DVCS is better
  equipped to record those details. As we mentioned before, it
  does have one disadvantage as compared to the file-scope
  approach: if a single file is separated from the distribution,
  the recipient won't see the contributors' copyright notices.
  But this can be easily remedied by including a single
  copyright notice in each file's header, pointing to the
  top-level file:

    Copyright YYYY-YYYY The Octave Project Developers

    See the COPYRIGHT file at the top-level directory
    of this distribution or at https://octave.org/COPYRIGHT.html.

followed by the usual GPL copyright statement.

For more background, see the discussion here:

  https://lists.gnu.org/archive/html/octave-maintainers/2020-01/msg00009.html


Most files in the following directories have been skipped intentinally
in this changeset:

  doc
  libgui/qterminal
  liboctave/external
  m4