Implement Enable property values "inactive" and "off" for uicontrol Edit (bug #57128)
* BaseControl.cc (updatePalette): Call setColor() with additional argument of
"QPalette::Active" and again with "QPalette::Inactive" so that colors are set
only for specific roles.  When Qt widget disabled it should now go gray.
* BaseControl.cc (init): Call 'setEnabled (! properties.enable_is ("off"))'
so that "Inactive" value still causes Qt Widget to be enabled.
* BaseControl.cc (update): Call 'setEnabled (! properties.enable_is ("off"))'
so that "Inactive" value still causes Qt Widget to be enabled.

* EditControl.cc (init): Check Octave "Enable" property and call Qt
setReadOnly() if value is "Inactive", but otherwise call Qt setEnabled() with
property value.
* EditControl.cc (updateSingleLine, updateMultiLine):  Check Octave "Enable"
property and call Qt setReadOnly() if value is "Inactive", but otherwise call
Qt setEnabled() with property value.