Return patch objects for bar,barh rather than hggroups when 'hist' is given (bug #59589).
* bar.m, barh.m: Reverse order of first two arguments to match new __bar__
prototype.

* __bar__.m: Reverse first two arguments of function and have the calling
function's name be the first argument and whether it is vertical/horizontal
as the second.  When returning patch objects, set CLIMMODE property to "auto".
When returning bar objects, set XLIMMODE property to "manual".
* __bar__.m (bars): Add new parameter "ishist" to function prototype.
Add new if clause which checks "ishist" and constructs patch objects for the
bars and then returns immediately.