allow more mixed-type assignments to work for sparse matrices (bug #60859)
This change allows indexed assignment to work when the LHS is a sparse
matrix (real or complex) and the RHS is a single precision scalar or
matrix (real or complex).

* ov-base-sparse.h, ov-base-sparse.cc (octave_base_sparse<T>::assign):
Declare RHS as a template type separate from the type of the sparse
array itself.  Move implementation to header file so that
instantiation will work automatically.

* Sparse.cc, Sparse.h (Sparse<T>::assign): Provide versions for scalar RHS.

* op-scm-cm.cc, op-scm-cs.cc, op-scm-m.cc, op-scm-s.cc, op-sm-cm.cc,
op-sm-cs.cc, op-sm-m.cc, op-sm-s.cc: Use DEFNDASSIGNOP_FN to define
functions for assignment.  Define and install mixed-type assignment
operations for float and float complex scalar and array assignments to
complex and double sparse arrays.

* test/sparse-assign.tst: New tests.
* test/module.mk: Update.