betaincinv.m: Overhaul function for performance and accuracy (bug #60539).
Switch algorithm from using 10 cycles of bisection followed by 20 rounds
of Newton's Method, to a pure Newton's Method that stops when accuracy
criteria of 1*eps(y) is reached.  Careful selection of initial starting
point to guarantee convergence is critical.

* betaincinv.m (bisection_method): Delete subfunction.
* betaincinv.m (newton_method): Remove stopping criteria based on max
iterations.  Remove fix for bug #60528 which is no longer necessary.
* betaincinv.m: New algorithm for selecting starting point for Newton's Method.
Documentation of method is in code.