don't emit signals for invalid objects in interpreter callbacks (bug #62863)
Some interpreter_event callbacks emit Qt signals.  The callback is
queued by the Octave interpreter and may be executed at
some arbitrary time.  It is possible that the object for which the
signal is intended to be executed may be deleted before the callback
executes.  If so, then the callback should not emit the signal.  In
some cases, execution of the entire callback function may be skipped.
In others, we still need to perform actions in the interpreter but
must skip the signals for Qt objects that no longer exist.

* command-widget.cc, file-editor-tab.cc, file-editor.cc,
octave-qscintilla.cc, main-window.cc, set-path-model.cc,
variable-editor-model.cc, variable-editor.cc: In all interpreter_event
callback functions that emit Qt signals, use a QPointer object to
protect the "this" pointer and check that it is still valid when the
callback executes.