Use data() instead of fortran_vec() where possible (bug #63803)
For all the instances of fortran_vec in libinterp/octave-value,
identify whether it can be treated as const, and replace
  T *tmp = foo.fortran_vec ();
with
  const T *tmp = foo.data ();
for those cases. This avoids the creation of a temporary copy,
saving memory and time.

Typically this pattern can be replaced when saving a variable to a
file, since the variable won't change at that point and can be
treated as const.