maint: Remove `#if 0` commented-out code from the codebase.
The codebase had 1850+ lines of code that were commented
out using the `#if 0` preprocessor directive. Running `hg annotate`
indicated that most of them had been commented out several years ago and
had not been edited in years, except maybe whitespace formatting,
leading to a large amount of commented-out code that was not
up-to-date with the rest of the codebase.

This edit removes almost all instances of `#if 0` from .h and .cpp files,
the only exception being a corner case where /* */ comments cannot be used
as they will interfere with BIST formatting.

If this removal proves to be too disruptive, please revert as required.