load-save.cc: Use temp file only when not in append mode (bug #63803)
load-save.cc in function load_save_system::save ():
Previously we were writing to a temp file so that if saving crashed we weren't
losing data already in the file from earlier.  That logic breaks down for append
mode, so this changeset does the temp file technique only for non-append mode.