avoid problems with $(shell ...) execution in Makefile
The $(shell ...) variable in GNU Make is expanded early before
Makefile variables are defined or exported.  This feature causes
trouble for some of Octave's build system shell scripts that may need
environment variables and that also need to be executed early to
generate lists of dependencies.  To work around this problem we can
use configure to substitute the needed variables instead of relying on
the Make environment.

See also the discussion here:

  https://octave.discourse.group/t/egrep-warning-egrep-is-obsolescent-using-ggrep-e/3484/37


* find-defun-files.in.sh: Rename from find-defun-files.sh.
* find-files-with-tests.in.sh: Rename from find-files-with-tests.sh.

* build-aux/module.mk: Update.

* configure.ac (AC_CONFIG_FILES): Include find-defun-files.sh and
find-files-with-tests.sh in the list of output files.

* libinterp/module.mk, liboctave/module.mk, scripts/module.mk: Remove
"$(srcdir)/" prefix from invocations or find-defun-files.sh and
find-files-with-tests.sh.