Improve Matlab compatibility of linspace for Inf endpoints.
Return an N-element vector if both START and END are the same.
When N is odd and [START, END] are [-Inf, Inf] return 0 for middle return value.
For all other elements besides START and END return NaN when one endpoint is
not finite.
* data.cc (Flinspace): Update and add new BIST tests for linspace.

* CRowVector.cc, dRowVector.cc, fCRowVector.cc, fRowVector.cc (linspace):
Add special case in input validation for START==END and return an N-element
vector of value END.  Check whether variable delta is Inf value and change
delta to NaN if it is.  For complex calculations, a separate isinf() check
is done for real and imaginary parts of delta.  For middle point when N
is odd, return 0 if START == -END, NaN if delta was NaN, or otherwise
(START + END) / 2.

* logspace.m: Update BIST tests.