perms.cc: Resequenced output for compatibility (bug #50426)
Previously, perms() was returning results in reverse lexicographic order.
For compatibility, it is required to do that only for inputs in ascending
order. Other inputs carry over that permutation to the results as well. This
edit changes the sort order, changes the docs, and fixes some BISTs to match.

Co-author: "Hendrik Koerner <koerhen@web.de>"
in https://savannah.gnu.org/bugs/index.php?63962#comment18