Fix converting short char arrays with invalid UTF-8 (bug #63930).
It looks like u8_conv_to_encoding checks the first input byte whether it looks
like an initial byte of a UTF-8 multi-byte surrogate. It returns an empty
string if the input buffer is too short to hold a surrogate of the expected
length. Additionally, it looks like it drops the trailing byte if it looks like
an initial byte of a multi-byte surrogate.
But we need it to "convert" the invalid byte sequences, e.g., to the
replacement character of the respective output encoding.

* liboctave/wrappers/uniconv-wrappers.c (octave_u8_conv_to_encoding_intern):
Add a new function that pads the input with zeros and zero terminates the
input string.
(octave_u8_conv_to_encoding, octave_u8_conv_to_encoding_strict,
octave_u8_conv_to_encoding_offsets): Use new function.

* scripts/strings/unicode2native.m: Add tests.