Don't use 2-digit exponent format for ticklabels when unnecessary.
* graphics.cc (calc_ticklabels): New boolean variable is_2digit_exp.
Cycle through ticklabel values using log10 to determine if maximum
exponent is greater than or equal to 10.  As soon as this condition
is found set is_2digit_exp to true and break out of loop.
Within ticklabel formatting loop, only print leading '0' if is_2digit_exp
is true and the current exponent to be printed is less than 10.