sparse-qr.cc: Fix compatibility with SuiteSparse 7.1.0 / CXSparse 4.0.4.
* liboctave/numeric/sparse-qr.cc (sparse_qr_rep::Q): "I" is a preprocessor
macro when <complex.h> is included. Rename variable to "I_mat".
(sparse_qr_rep::tall_solve, sparse_qr_rep::wide_solve): "cs_complex_t" is
"double _Complex" in SuiteSparse 7.1.0 / CXSparse 4.0.4 or newer.  So, it is
not possible to use it as a constructor with two arguments.  It was
"std::complex<double>" in previous versions.  Use the default one-argument
constructor that works with both instead.