Remove configuration option `--disable-vm-evaluator`
The `--disable-vm-evaluator` flag was being used only to selectively
allow/disallow the `make check-vm` option, but the VM was being
built and could be used whether that flag was passed or not.
This changeset removes that flag, allowing the VM to always be
built and available to the end user. To anyone building Octave
from source, this makes both `make check` and `make check-vm`
available. (The VM is still disabled when Octave starts, and the
user must explicitly invoke it with `__vm_enable__ (1)`.)

* configure.ac: Remove flag and corresponding variables.
* Makefile.am, test/Makefile.am: Remove use of flag,
allow `check-vm` to be available always.
* compile.cc: Edit FIXME and docstring to not mention the flag.