doc: Placate Texinfo warnings
Texinfo gave the following warnings:

* @verb{} should not be used in @item{} or any other construct that
generates a reference.

* @flushright and @image should not be on the same line.

* For the section "Hardware vs. Software Rendering", the section title
cannot have a dot, which would interfere with link generation.

This patch makes the following changes:

Replace @verb{} by @qcode{} and verify that it does not interfere
with info's formatting, though it may have done so in the past.

Move @image after @flushright to a separate line.

Change "vs." to "vs".