avoid warning about ignored return value
Recent versions of Qt tag the QTranslator::load with the [[nodiscard]]
attribute so that compilers are encouraged to warn if the return value
is not used.  Avoid that possible compiler warning for our uses of
QTranslator::load and also warn at runtime if translation files fail
to load properly.

* gui-settings.h, gui-settings.cc (gui_settings::load_translator):
New functions.
(gui_settings::config_translators): Use them instead of calling
QTraslator::load directly.