test: speed up BIST for the central part of convn with 'full' shape.
The test for bug #39314 requires a particularly long array.  That array has
176,800 elements (size of [13 17 8 10 10]).  The test checks if the central
part of a "full" convolution is the same as the "valid" convolution (it is not,
this is bug #39314).  The assert command then shows all the elements that are
different (effectively all) which makes it extremely slow for such a large
array.  This commit changes the test so it does the comparison first, and only
passes a true/false to "assert".

The end performance improvement will be machine and options dependent but in my
machine it reduces the test time from 6 to 0.7 seconds.  On Discourse
(https://octave.discourse.group/t/conv2-test-time-rather-long/5446/18) some
people were experiencing 15 minutes for this test alone.

* conv2.cc: Check equality of observed and expected arrays and pass true/false
result to assert.