GitHub-CI (cygwin): Disable visibility attributes for Cygwin.
* .github/workflows/make.yaml (cygwin): None of the explicitly instantiated
template classes are exported from the libraries currently. This might be an
issue with the way we are instantiating the template classes. Or it might be an
issue with the compiler of Cygwin (currently GCC 11.4.0).
Disable building with visibility attributes in the Cygwin CI-rules for now.
Maybe, test again when Cygwin updated to a newer major version of GCC.