print_usage.m: Speed up Texinfo processing for common cases
Previously, `print_usage` was always calling `__makeinfo__` which in turn
calls `system ("makeinfo")`. This turned out to be time-consuming, especially
for `make check`, where profiling showed that a majority of the time was being
spent in `system ("makeinfo")`. Discussion:
https://octave.discourse.group/t/calls-to-system-slowing-down-make-check/5563

This patch converts the most commonly occurring `print_usage` Texinfo commands
to plain text without requiring the full makeinfo. As a consequence,
`make check` is sped up by 2X to 2.5X on Linux and by 5X to 10X on Windows.

* print_usage.m:
** New function `__makeinfo_restricted__` to process specific Texinfo commands.
** Use that function in `get_usage_texinfo` and fall back to `__makeinfo__`
   if it fails.

* movfun.m: Replace a nested `@var` occurrence that caused problems for the
new code.