colormap.m: Override colormaps on axes objects for Matlab Compatibility (bug #65674).
* colormap.m: Update documentation to mention that first input arg may be a figure.
Redo input validation to accept either an axes or figure object for the target of
the function.  For Matlab compatibility, if the target is a figure then also find
any axes object of the target and delete any colormaps present.  Add BIST test for
bug #65674.  Reorder BIST %!error tests to follow their order in the code.