Return early if sort dimension exceeds array dimension (bug #65712)
It turn out that calling `sort (A, dim)` where dim is greater than `ndims (A)`
was creating a larger temp array to accommodate dim, so `sort (A, 1e9)` was
transiently using 8GB, while `sort (A, inf)` was causing either OOM or
a segfault, depending on the compiler and its settings.

This patch changes the internal code for array sorting so that when the sort
dimension exceeds the array's dimensions, the array is returned unchanged.
This is analogous to `all (true(3,3), 100)` and similar expressions.

* Array-base.cc (in function `Array<T, Alloc>::sort (int, sortmode)`):
Copy array and return early if sort dimension exceeds array dimension.

* data.cc (in Fsort): Add BISTs.