oct-binmap.h: Remove FIXME notes after testing
oct-binmap.h: A couple of FIXME notes mentioned the idea of keeping track of
whether fcn call results in a 0, and potentially skipping maybe_compress()
if no new zeros are created. That idea was tested out for some 10 different
input combinations and found to make no performance difference: the time for
a binary operation between a scalar and a 1000x1000 matrix (in each order,
tested 1000 times each) changed by less than 3 microseconds, and sometimes
took that much *longer* than not keeping track of new zeros.

Statistically it made no difference, so the simpler code is preferred
and the FIXME notes have been removed.