liboctmex: Check if hg id matches other Octave libraries.
* libmex/liboctmex-build-info.h, libmex/liboctmex-build-info.in.cc: Add new
files that export a function that returns the hg id of the repository at the
time of building the library.
* libmex/module.mk: Add new files to build system.
* src/main-cli.cc (check_hg_versions), src/main-gui.cc (check_hg_versions):
Compare hg id of liboctmex with other libraries. This also means that the
liboctmex library is loaded early on when starting Octave. That helps to avoid
potential issues with loading the "wrong" version of the library later on when
an .oct file library is loaded.
* src/module.mk: Link main programs to liboctmex.