mkoctfile: Use more resilient approach to link Octave libraries.
* configure.ac: Add path to Octave libraries to linker search path in
MKOCTFILE_OCT_LINK_DEPS when linking to these libraries (i.e., when linking all
dependencies or on Windows). That is more resilient than adding it to LDFLAGS in
mkoctfile because LDFLAGS could be overridden with the value of an environment
variable with that name. Expand configure variables in MKOCTFILE_OCT_LINK_DEPS.
(The latter was already missing since 0bf6b7e1367c.)
* src/mkoctfile.in.cc: Remove OS specific code that is no longer needed.