Check if major SOVERSION of liboctmex matches when loading .mex file libraries.
* Makefile.am: Add variable with major SOVERSION of liboctmex. Its value can be
used by the build rules for multiple libraries of Octave.
* libmex/module.mk: Use major SOVERSION of liboctmex as defined in Makefile.am.
* libmex/mex.h: Add symbol that contains the major SOVERSION of liboctmex in
each .mex file library. (All .mex file libraries need to include this header.)
* src/mkoctfile.in.cc (main): Add preprocessor flag to define
OCTAVE_MEX_SOVERSION with the correct value when building .mex files.
* src/module.mk: Add preprocessor flag to define OCTAVE_MEX_SOVERSION  with the
correct value when building "mkoctfile" executable.
* libinterp/corefcn/dynamic-ld.cc (dynamic_loader::load_mex): Check if the
SOVERSION of the liboctmex library that was used when building the .mex file
library matches the SOVERSION of the liboctmex library that is used by the
running version of Octave.
* libinterp/corefcn/module.mk: Add preprocessor flag to define
OCTAVE_MEX_SOVERSION with the correct value when building the libcorefcn object
library.