plot3: Draw marker when only a single point is plotted.
Although this is not Matlab-compatible, it is very confusing to
issue a plot command and have nothing drawn.  Octave had already
made the choice for the 2-D plot command to draw a marker for
a single point so now the 2-D and 3-D cases behave the same.

* NEWS.10.md: Announce change in Graphics Backend section.

* plot3.m: Check input data for being a single point and set 'marker' to '.'
and 'linestyle' to '-'.