Modify BIST tests for password functions to pass on all platforms (bug #67379)
On OpenBSD using endpwent() does not reset file pointer on password database
to first entry.  Instead, use setpwent() which reliably works on all OS
to set the file pointer to first entry.

* system.tst: Add HAVE_GETPWENT, HAVE_SETPWENT, and HAVE_ENDPWENT as necessary
to %!testif header so that BIST tests are guaranteed to have the functions they
need to run.  Replace use of endpwent() to position password file pointer with
setpwent().