Do not treat struct as return argument type for saveobj differently (bug #45833).
* libinterp/octave-value/ov-classdef.cc (octave_classdef::saveobj): Set
indicator for a different type of the return argument of an overloaded "saveobj"
method anytime the type and class does not match the original type and class of
the object that is to be saved.
(octave_classdef::loadobj): Adjust for the change in octave_classdef::saveobj.
Emit more specific warnings for (currently) unsupported scenarios.
* test/classdef-load-save/classdef-load-save.tst: Adjust test for saveobj
returning a structure to the actually expected behavior. Add test for saveobj
returning an object (of the correct class).
* test/classdef-load-save/saveobj_obj_class.m: Add definition for class that
returns an object of the correct class from the overloaded "saveobj" method and
that does not overload "loadobj".
* test/classdef-load-save/saveobj_struct_class.m: Rename from "saveobj_class.m".
* test/classdef-load-save/module.mk: Add new and renamed files to dist tarball.