fzero.m: Improve accuracy of final result (bug #67407)
Algorithm produces bracketing estimate of x with low-estimate in variable 'a'
and high-estimate in variable 'b'.  Choose return value x to be 'a' or 'b'
based on which has the smaller function value (i.e., is closer to zero).

* fzero.m: Assign output 'x' and 'fval' based on which value of bracket
('a' or 'b') has smaller absolute value fcn(a) or fcn(b).

* NEWS.11.md: Note improved accuracy in General Improvements section.