fill.m: Overhaul input validation to be Matlab-compatible (bug #67535)
* NEWS.11.md: Announce changes in Matlab compatibility section.

* fill.m: Document that specifying one color per polygon can be done with any
vector (not just row vector).  Use numel() in preference to length().  Add
FIXME note about missing validation that inputs X, Y, and C are 2-D.  Replace
input logic for vector/matrix combinations with longer code that attempts to
match the length of the vector to one of the dimensions of the matrix data.
Replace use of isequal() (slow) with faster tests that don't require function
call.  Validate that RGB triplet data is in the range [0,1].