Add "Source" and "EventName" fields to KeyEvent struct for Matlab Compatibility (bug #44900)
* NEWS.11.md: Announce addition in Matlab Compatibility section.

* BaseControl.cc (eventFilter): Rename variable "keyData" to "keyEvent".
Use setfield() to add "Source" and "EventName" fields for UIControl objects.

* Canvas.cc (canvasKeyPressEvent, canvassKeyReleaseEvent):
Use setfield() to add "Source" and "EventName" fields for figure objects.

* Table.cc (eventFilter): Rename variable "keyData" to "keyEvent".
Use setfield() to add "Source" and "EventName" fields for figure objects.

* __init_fltk__.cc (handle): Use assign() to add "Source" and "EventName"
fields to keyEvent structure.  Add FIXME note that "Source" is empty and
more code would be needed to get this to work.

* genpropdoc.m: Update documentation for "KeyPressFcn" and "KeyReleaseFcn" for
Figure, UIControl, and UITable objects.