inputdlg.m: Overhaul function for Matlab compatibility (bug #49143)
* inputdlg.m: Rewrite docstring.  Specifically document 1x2 row vector form of
ROWSCOLS spec.  Replace narginchk() with if (...) and call to print_usage().
Move input validation for ROWSCOLS to be adjacent to where input is gathered.
Code exception case for 1x2 row vector form of ROWSCOLS spec.  Reverse sense of
if () tests to have error() calls occur higher in the code.  Rewrite and add
BIST tests.