Shorten code in filter.cc (bug #61674)
Previously, the filtering calculations were split across an outer loop and
an inner loop to allow for both performance and Ctrl-C to be used. It was found
recently that Ctrl-C now works even without the outer and inner loop mechanism,
so this changeset simplifies that part of the code.

* filter.cc: Merge outer and inner loops into a single loop. Rewrite comments
to match code. Define variable y only after input validation is complete.
Define variable 'zero' earlier and use it in one more place.