mean.m: No longer accept char inputs.
No longer accept char inputs for Matlab compatibility and for consistency with
other Octave statistics functions.
Clean up code.

* NEWS.11.md: Announce change.

* mean.m: Move input validation for input X to top of function and check for
isnumeric() and islogical().  Remove input validation code for OUTTYPE "char".
Rename variable "idx" to "nanx" for clarity.  Use isindex() for input
validation of VECDIM rather than handrolled code.  New variable "dim" for
code readability which is an alias for VECDIM when VECDIM is a scalar.
Remove BIST tests for char input.  Re-order BIST tests to occur in the order
in which the errors are generated in the code.