pkg: Rework downloading package metadata
Passing the '-refresh' option to 'pkg update' etc was previously causing a fresh
download of metadata from the server for each installed package. This changeset
fixes that behavior so that passing '-refresh' only causes one new download
of metadata, irrespective of the number of installed packages.

All changes in directory 'scripts/pkg/':

* pkg.m: Preload package metadata, call get_validated_pkg_list() directly.

* private/get_pkg_info.m: Remove function calls to update package metadata,
  change calling form, update documentation to match.

* private/get_validated_pkg_list.m: Rework logic to save metadata to cache.

* private/search_packages.m: Change calling form, update documentation to match.