pkg: Focus on search terms in search results
Previously, when using "pkg search", the search result would show up to the
first 80 characters of the package description. For example, running the command
"pkg search statistics" would show one of its results like this:

stk |  2.8.1  | The STK is a (not so) Small Toolbox for Kriging.  Its primary focus is on the...

The minor shortcoming of such a presentation is that it is not clear why that
package is being returned as a result for the search term "statistics".

This changeset "zooms in" on the actual search terms used, resolving a FIXME,
resulting in this result for the same command:

stk |  2.8.1  | ...as Geostatistics, Machine Learning, Non-parametric Regression, etc.).

The hope is that this is more useful to the user to decide what package is more
relevant to their search context.

* scripts/pkg/private/search_packages.m: Make the above change.