Create new console when creating child processes on Windows (bug #67794).
When creating processes with `_wsystem` on Windows, the child process shares
the console with the parent process. To avoid desynchronization between the
command window widget and the console used by the parent process, start child
processes with their own (hidden) console.

* liboctave/system/oct-sysdep.cc (octave::sys:system): On Windows, create child
processes with a new console (to be prepared in case the console of the parent
process was created with ALLOC_CONSOLE_MODE_NO_WINDOW which is the case when
using the GUI in Windows 11 24H2 or later).

* libinterp/corefcn/toplev.cc (Fsystem): Remove special case for Windows that
is no longer necessary (and would cause issues with the new implementation of
octave::sys:system).