Create new console for pipe to child processes on Windows (bug #67794).
When creating pipes to a new process with `_wpopen` on Windows, the child
process shares the console with the parent process. To avoid desynchronization
between the command window widget and the console used by the parent process,
start child processes with their own (hidden) console.

* libinterp/corefcn/sysdep.cc (octave::popen): On Windows, create child
processes for a pipe with a new hidden console.
(octave::pclose): Close handle to child process when closing pipe.