maint: Relocate BIST tests for sparse matrix broadcasting
All tests for sparse matrix broadcasting previously found at the bottom of
CSparse.cc and dSparse.cc files are relocated to a new dedicated file in
test directory, or to the files with the corresponding functions.

* sparse-broadcasting.tst: New file.  All broadcasting tests related to math,
comparison, and boolean operations regarding sparse matrices are located in
this file.

* test/Makefile.am: Add file sparse-broadcasting.tst to build system.

* data.cc (Fcumprod, Fcumsum, Fprod, Fsum, Fsumsq): Add related tests to their
corresponding functions.

* max.cc (Fmin, Fmax) : Add related tests to their corresponding functions.